merge

+++ src/main/java/insa/rennes/BatchJob.java
@@ -48,7 +48,7 @@ public class BatchJob {
 
 		fifaRanks = env.readCsvFile(fifaRanksPath)
 				.ignoreFirstLine()
-				.types(Integer.class, String.class, Float.class, Float.class, Integer.class, String.class)
+				.types(Integer.class, String.class, Float.class, Integer.class, Integer.class, String.class)
 				.flatMap(new FifaRankingDateConverter())
 				.flatMap(new FifaRankingStats())
 				.groupBy(0)
@@ -104,9 +104,9 @@ public class BatchJob {
 		}
 	}
 
-	public static class FifaRankingStats implements FlatMapFunction<Tuple6<Integer, String, Float, Float, Integer, Date>, Tuple3<String, Integer, Integer>> {
+	public static class FifaRankingStats implements FlatMapFunction<Tuple6<Integer, String, Float, Integer, Integer, Date>, Tuple3<String, Integer, Integer>> {
 		@Override
-		public void flatMap(Tuple6<Integer, String, Float, Float, Integer, Date> in, Collector<Tuple3<String, Integer, Integer>> out) throws Exception {
+		public void flatMap(Tuple6<Integer, String, Float, Integer, Integer, Date> in, Collector<Tuple3<String, Integer, Integer>> out) throws Exception {
 			out.collect(new Tuple3(in.f1, in.f0, 1));
 		}
 	}
